+++ example/index.ts
@@ -2,8 +2,13 @@ import { Memoirist } from '../src'
 
 const router = new Memoirist()
 
-router.add('GET', '/api/search/:term', '/api/search/:term')
-router.add('GET', '/api/abc/view/:id', '/api/abc/view/:id')
-router.add('GET', '/api/abc/:type', '/api/abc/:type')
+router.add(
+    'GET',
+    '/api/search/:term?/:name?/:age?',
+    '/api/search/:term?/:name?/:age?'
+)
 
-console.log('GOT', router.find('GET', '/api/search/type'))
+// router.add('GET', '/api/search/:term?/:name?/:age?', 'a')
+// console.log(router.history.map(([method, path, store]) => path))
+
+// console.log('GOT', router.find('GET', '/api/search/term/name/age/10'))